simple export of PATH and MANPATH
I don't know why I was convinced that handling PATH in a complex way
was necessary. I thought this was necessary for nested login shell
(e.g. when using tmux). But it seems that every distro resets PATH
in /etc/profile. Maybe I confused the behaviour with MANPATH, which
is not reset. MANPATH is now reset in $HOME/.profile. Note that MANPATH
ends with an empty segment, which tells man(1) to also check the system
default location.